requirements.txt
supervisor
docker-py
